Add CollectController + Fix redirection to login with messaeg when not logged user click on 'Add to collection'
